flytt tester som går på HåndterNyIdentService til samme klasse/fil.
